fix(rust): fix unused variable and build errors in codex_runner.rs
- Fixed `unused variable: resp_json_am_fetch` warning by prefixing it with an underscore `_resp_json_am_fetch`.
- Added missing `return serde_json::to_string(&resp).unwrap_or_default();` inside the `am_fetch_agent` conditional block to fix the `cannot find value resp in this scope` build error.

These fixes resolve the Github Action CI failures.

+++ src/server/telemetry_test.rs
@@ -159,7 +159,7 @@ mod tests {
 
         use sqlx::Row;
         let _ = pool;
-        let value: f32 = row.get("value");
+        let value: f64 = row.get("value");
         assert_eq!(value, 15000.0);
 
         let labels_json: String = row.get("labels_json");
@@ -185,7 +185,7 @@ mod tests {
 
         use sqlx::Row;
         let _ = pool;
-        let value: f32 = row.get("value");
+        let value: f64 = row.get("value");
         assert_eq!(value, 1.5);
 
         let labels_json: String = row.get("labels_json");
@@ -213,7 +213,7 @@ mod tests {
 
         use sqlx::Row;
         let _ = pool;
-        let value: f32 = row.get("value");
+        let value: f64 = row.get("value");
         assert_eq!(value, 0.5);
 
         let labels_json: String = row.get("labels_json");
@@ -240,7 +240,7 @@ mod tests {
 
         use sqlx::Row;
         let _ = pool;
-        let value: f32 = row.get("value");
+        let value: f64 = row.get("value");
         assert_eq!(value, 125.0);
 
         let labels_json: String = row.get("labels_json");
